+++ src/main/resources/static/js/comm/path/pathList.js
@@ -14,8 +14,16 @@ document.addEventListener('DOMContentLoaded', function() {
 	
 	document.getElementById('btnWrite').addEventListener('click', function() {
 		if (!memId || memId == 'anonymousUser') {
-			sessionStorage.setItem("redirectUrl", location.href);
-			location.href = "/login";
+			showConfirm("로그인 후 이용 가능합니다.", "로그인하시겠습니까?",
+				() => {
+					sessionStorage.setItem("redirectUrl", location.href);
+					location.href = "/login";
+				},
+				() => {
+
+				}
+			);
+
 		} else {
 			location.href = "/comm/path/pathInsert.do";
 		}
