+++ src/main/resources/static/js/include/chatModal.js
@@ -297,8 +297,15 @@ function closeChatModal(){
 // 모달 열기
 async function openChatModal(){
 	if(!memId || memId=='anonymousUser') {
-		sessionStorage.setItem("redirectUrl", location.href);
-		location.href = "/login";
+		showConfirm("로그인 후 이용 가능합니다.", "로그인하시겠습니까?",
+			() => {
+				sessionStorage.setItem("redirectUrl", location.href);
+				location.href = "/login";
+			},
+			() => {
+
+			}
+		);
 	} else {
 		axios.post("/admin/las/chatVisitLog.do");
 		await printChatRoomList();
